render ply as pcd if specified
