+++ app/packages/looker-3d/src/fo3d/mesh/Ply.tsx
@@ -1,15 +1,17 @@
 import { getSampleSrc } from "@fiftyone/state";
 import { useLoader } from "@react-three/fiber";
-import { useEffect, useMemo, useState } from "react";
-import { BufferGeometry, Mesh, Quaternion, Vector3 } from "three";
+import { useEffect, useMemo, useRef, useState } from "react";
+import { BufferGeometry, Mesh, Points, Quaternion, Vector3 } from "three";
 import { PLYLoader } from "three/examples/jsm/loaders/PLYLoader";
 import {
   FoMeshBasicMaterialProps,
   FoMeshMaterial,
+  FoPointcloudMaterialProps,
   PlyAsset,
 } from "../../hooks";
 import { useMeshMaterialControls } from "../../hooks/use-mesh-material-controls";
 import { useFo3dContext } from "../context";
+import { usePcdMaterial } from "../point-cloud/use-pcd-material";
 import { getBasePathForTextures, getResolvedUrlForFo3dAsset } from "../utils";
 
 interface PlyProps {
@@ -20,6 +22,45 @@ interface PlyProps {
   scale: Vector3;
   children?: React.ReactNode;
 }
+
+const PlyWithPointsMaterial = ({
+  name,
+  geometry,
+  defaultMaterial,
+}: {
+  name: string;
+  geometry: BufferGeometry;
+  defaultMaterial: FoMeshMaterial;
+}) => {
+  const overrideMaterial = {
+    shadingMode: "height",
+    customColor: defaultMaterial["color"] ?? "#ffffff",
+    pointSize: 2,
+    attenuateByDistance: false,
+    opacity: defaultMaterial.opacity,
+  } as FoPointcloudMaterialProps;
+
+  const pointsContainerRef = useRef();
+
+  const pointsMaterialJsx = usePcdMaterial(
+    name,
+    geometry,
+    overrideMaterial,
+    pointsContainerRef
+  );
+
+  const mesh = useMemo(() => new Points(geometry), [geometry]);
+
+  if (!geometry || !pointsMaterialJsx) {
+    return null;
+  }
+
+  return (
+    <primitive ref={pointsContainerRef} object={mesh}>
+      {pointsMaterialJsx}
+    </primitive>
+  );
+};
 
 const PlyWithMaterialOverride = ({
   name,
@@ -77,7 +118,7 @@ const PlyWithNoMaterialOverride = ({
 
 export const Ply = ({
   name,
-  ply: { plyPath, preTransformedPlyPath, defaultMaterial },
+  ply: { plyPath, preTransformedPlyPath, defaultMaterial, isPcd },
   position,
   quaternion,
   scale,
@@ -129,6 +170,16 @@ export const Ply = ({
       return null;
     }
 
+    if (isPcd) {
+      return (
+        <PlyWithPointsMaterial
+          name={name}
+          geometry={geometry}
+          defaultMaterial={defaultMaterial}
+        />
+      );
+    }
+
     if (isUsingVertexColors) {
       return (
         <PlyWithMaterialOverride
@@ -150,6 +201,7 @@ export const Ply = ({
     isGeometryResolved,
     isUsingVertexColors,
     geometry,
+    isPcd,
     name,
     defaultMaterial,
   ]);
